refactor(mobile): split ChatMessage.tsx 533->146 LOC (Cluster A2)
CLAUDE.md mandates components <400 LOC. ChatMessage.tsx was 533. Extracted
into co-located chat/ modules, behavior-preserving:
- message-types.ts: the message domain types.
- message-content.ts: the pure markdown parser (parseMessageContent /
  parseInlineCode) + MONO_FONT_FAMILY — now unit-testable.
- useCopyWithClear.ts: DRY'd the duplicated clipboard copy + 30s
  auto-clear (SEC-MOB-001) logic shared by the two code blocks.
- CodeBlock.tsx / ThinkingBlock.tsx / TextBlock.tsx: the three sub-renderers.
ChatMessage.tsx keeps the main component + re-exports the full public API
(types + parseMessageContent), so every existing importer of './ChatMessage'
is unchanged.

The extracted code blocks are now in their own files with raw Pressables, so
they got accessibilityRole/Label (required + verified by the C3 a11y guard)
- the originals lacked them on the copy/expand buttons.

Tests: +10 for the now-extracted pure parser (fences, inline code, the
backticks-inside-a-fence edge case). Mobile gate green: typecheck, lint,
1715 jest tests (full suite — every ChatMessage consumer still passes, so
behavior is preserved).
